Viewports and Footers
 I added and styled the footer for the website. I originally wanted to make changes for the navbar and script it to be sticky, but I instead I went to the footer for time and preparation reasons.

I made the footer to be similar to https://rewindapp.flatstudio.co/  's footer with logos on it. Using the pixel sprites as placeholder images for my Github, Twiiter, Facebook, and Linkdin.

The font styling needs more work. I wanna change the h1 h2 h3 and h4 fonts to be condensed instead of just being a bold version of the roboto font.

going to leave the footer as is for now, with the intention of tweaking the fonts, the logo links, and the dividers to be centered later on.

Next commit will be to make the navbar sticky and responsive