comments and suggestions push
I was reviewing my html and styling and decided I needed to change a few things.

1. I changed the bg color of .nav-bar to be bgColor3 because bgColor1 was disonant to the color scheme
2. I added Comments and suggestions to the HTML tags for how I should style them to look better or more efficient for the user experience or aesthetics (both animations and static looks)
3. Added dividers and labels for the arranged CSS to make it easier to read and locate the stylings for the the style of the webpages

P.S. using this arragned method for CSS is so much better. I plan to replace all the css in my project to use SASS instead of base CSS to make the styling for each element nested instead of just laid out in a big pile/column of CSS.