wirings, restructurings, and minor changes
wired many of the loose ends together so that everything is in order. (basically put urls to the links of each page)

changed the about section in the home page to be better fitting

changed the gallery in the portfolio section to be an actual useful gallery

changed the color scheme so that it's better to look at. although it's big shift in aesthetics to me, it's a better read for my eyes and better colors

(also added a code archive for disused snippets of html and css)